add ability to view attempt as user from report
This adds a column "View attempt" to the SCORM basic report. It contains links to view each attempt in review mode. To make this possible, instances of $USER in most places in the plugin had to be replaced with a new variable $viewing_user, which is set based on the `userid` parameter in the URL.
Only site admins can do this. It might be worth adding a capability instead, so teachers can be given the permission.