don't add an item to history when redirecting to SCORM
loadSCO.php used to use `location = <url>` to redirect to the SCORM activity's page. If this happens too late after the page loads (I think, maybe it's browser-dependent) then the loadSCO page gets added to the browser's history.
If the student navigates backwards from the SCORM activity, the iframe goes back to loadSCO, and they're redirected on to the SCORM activity again, except this time the API has already been initialised.

This commit changes the relevant lines to use `location.replace(url)` instead, so that no history item is added, and backwards navigation takes you back to view.php